+++ test/esockd_access_tests.erl
@@ -44,10 +44,10 @@ mask_test() ->
     ?assertEqual(16#FFFFFF80, mask(25)).
 
 range_test() ->
-    {ok, Start, End} = range("192.168.1.0/24"),
+    {Start, End} = range("192.168.1.0/24"),
     ?assertEqual({192,168,1,0}, itoa(Start)),
     ?assertEqual({192,168,1,255}, itoa(End)),
-    {ok, Start1, End1} = range("10.10.0.0/16"),
+    {Start1, End1} = range("10.10.0.0/16"),
     ?assertEqual({10,10,0,0}, itoa(Start1)),
     ?assertEqual({10,10,255,255}, itoa(End1)).